endian fix for dataset and swizzled dataset

+++ fitsToHDF5_IDIA.py
@@ -147,13 +147,13 @@ class Dataset:
         
         axis = self.swizzle_axis_numeric(axis_name)
         swizzled = self.hdu_group.require_group("SwizzledData")
-        swizzled.create_dataset(axis_name, data=np.transpose(self.data, axis))
+        swizzled.create_dataset(axis_name, dtype='f4', data=np.transpose(self.data, axis))
     
     def write(self, args):
         # write this dataset
         # TODO TODO TODO check that the number of chunks matches the data dimensions
         logging.info("Writing main dataset...")
-        self.hdu_group.create_dataset(self.name, data=self.data, chunks=tuple(args.chunks) if args.chunks else None)
+        self.hdu_group.create_dataset(self.name, dtype='f4', data=self.data, chunks=tuple(args.chunks) if args.chunks else None)
         
         # write statistics
         for axis_name in args.statistics:
